requirements.txt
Django==6.0.2
djangorestframework==3.16.1
psycopg[binary]==3.3.3
django-cors-headers>=4.9.0
djangorestframework-simplejwt==5.5.1
Pillow
python-decouple
bleach